Fix ICAST_IDIV_CAST_TO_DOUBLE SpotBugs warning in StatisticalTests
This commit fixes the integer division cast to double issue reported by SpotBugs. The issue was that expressions like ((4 * i + 1) * z) / sqrt(bitString.length()) were performing integer multiplication first, which could lead to overflow or precision loss.

The fix is to cast the integer variable 'z' to double before multiplication, ensuring floating-point arithmetic throughout the calculation.

+++ TLS-Scanner-Core/src/main/java/de/rub/nds/tlsscanner/core/vector/statistics/StatisticalTests.java
@@ -100,11 +100,11 @@ public class StatisticalTests {
             probSum1 =
                     probSum1
                             + dst.cumulativeProbability(
-                                    ((4 * i + 1) * z) / sqrt(bitString.length()));
+                                    ((4 * i + 1) * (double) z) / sqrt(bitString.length()));
             probSum1 =
                     probSum1
                             - dst.cumulativeProbability(
-                                    ((4 * i - 1) * z) / sqrt(bitString.length()));
+                                    ((4 * i - 1) * (double) z) / sqrt(bitString.length()));
         }
 
         sumStart = ((-bitString.length() / z) - 3) / 4;
@@ -114,11 +114,11 @@ public class StatisticalTests {
             probSum2 =
                     probSum2
                             + dst.cumulativeProbability(
-                                    ((4 * i + 3) * z) / sqrt(bitString.length()));
+                                    ((4 * i + 3) * (double) z) / sqrt(bitString.length()));
             probSum2 =
                     probSum2
                             - dst.cumulativeProbability(
-                                    ((4 * i + 1) * z) / sqrt(bitString.length()));
+                                    ((4 * i + 1) * (double) z) / sqrt(bitString.length()));
         }
 
         pValue = 1 - probSum1 + probSum2;
